added tests for input redirection, also some implementation
