+++ Makefile
@@ -44,7 +44,7 @@ COMMAND_EXPANSION_NAMES = substitute_envvars.c expand.c
 COMMAND_EXPANSION_F = command_expansion
 COMMAND_EXPANSION_SRCS = $(addprefix $(COMMAND_EXPANSION_F)/, $(COMMAND_EXPANSION_NAMES))
 
-EXECUTION_NAMES = exec_text_tree.c exec_text_tree_node.c exec_text_tree_controls.c redirections.c pipes.c files.c is_pipe.c is_redir.c
+EXECUTION_NAMES = exec_text_tree.c exec_text_tree_node.c exec_text_tree_controls.c redirections.c pipes.c files.c is_pipe.c is_redir.c is_redir_file.c navigation.c
 EXECUTION_F = execution
 EXECUTION_SRCS = $(addprefix $(EXECUTION_F)/, $(EXECUTION_NAMES))
 
@@ -58,7 +58,7 @@ ENDPOINT_NAME = main.c
 SRC_SRCS = $(addprefix $(SOURCE_F)/, $(SRC_NAMES))
 ENDPOINT_SRC = $(addprefix $(SOURCE_F)/, $(ENDPOINT_NAME))
 
-TEST_NAMES = input_to_text_tree_test.c ft_split_str_test.c tokenize_cmd_test.c unit_tests.c e2e_tests.c expand_tree_test.c get_envvars_test.c
+TEST_NAMES = input_to_text_tree_test.c ft_split_str_test.c tokenize_cmd_test.c unit_tests.c e2e_tests.c expand_tree_test.c get_envvars_test.c e2e_comparing_functions.c e2e_utils.c
 TEST_ENDPOINT_NAME = main_test.c
 TEST_SRCS = $(addprefix $(TEST_F)/, $(TEST_NAMES))
 TEST_ENDPOINT_SRC = $(addprefix $(TEST_F)/, $(TEST_ENDPOINT_NAME))
@@ -186,13 +186,16 @@ tania: $(OBJ_DIRS) $(TANIA_OBJ_F) $(OBJS) $(TANIA_ENDPOINT_OBJ)
 all_trapped:
 	$(PREFIX)make PREPROC_DEFINES="$(PREPROC_DEFINES) -DFT_CALLOC_IF_TRAPPED -DFANCY_IFACE" all
 
+all_fancy:
+	$(PREFIX)make PREPROC_DEFINES="$(PREPROC_DEFINES) -DFANCY_IFACE" all
+
 vania:
 	$(PREFIX)cd libft && make fulltest_trapped_nonorm
 	$(PREFIX)make fclean testfclean
 	$(PREFIX)make minivania
 
 minivania:
-	$(PREFIX)make all test && ./$(TEST_FNAME)
+	$(PREFIX)make all_fancy test && ./$(TEST_FNAME)
 run:
 	$(PREFIX)./minishell --debug 136 --command $(CMD)
 debug:
