+++ sources/execution/execution_internal.h
@@ -6,7 +6,7 @@
 /*   By: inikulin <inikulin@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/13 17:46:39 by inikulin          #+#    #+#             */
-/*   Updated: 2025/01/12 14:11:39 by inikulin         ###   ########.fr       */
+/*   Updated: 2025/01/13 18:23:16 by inikulin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,6 +33,15 @@ typedef struct s_executor
 
 # define IN 0
 # define OUT 1
+# define TYPE 2
+# define COMMAND 1
+# define TO_PIPE 2
+# define FROM_PIPE 4
+# define TO_OUT_FILE 8
+# define FROM_IN_FILE 16
+# define OUT_FILE 32
+# define IN_FILE 64
+# define IGNORED_FILE 128
 
 t_executor	*make_executor(t_treenode *node, t_param *param);
 int			executor_finalize(t_executor *e);
@@ -40,21 +49,28 @@ int			execute_node(t_executor *executor);
 int			execute_text_tree_node(t_executor *executor);
 int			exec_rec(t_executor *e);
 int			redirections(t_executor *e);
-int			takes_part_in_pipe(t_treenode *node);
-int			setup_file(t_executor *e, t_treenode *node, int i);
-int			close_fds(t_executor *e, int tgt);
-int			setup_pipe(t_executor *e, int i);
-int			fd_ok(int fd);
-int			scroll_chain(t_executor *e, int tgt);
 
+int			setup_out_file(t_executor *e, t_treenode *node, int i);
+int			setup_in_file(t_executor *e, t_treenode *node, int i);
+int			setup_pipe(t_executor *e, t_treenode *node, int i);
+
+int			fd_ok(int fd);
+int			close_fds(t_executor *e, int tgt);
+int			scroll_chain(t_executor *e, int tgt);
+int			rollback_input_files_fds(t_executor *e, t_treenode *node, int i);
+
+int			takes_part_in_pipe(t_treenode *node);
 int			is_pipe_or_redir(char *s);
 int			is_pipe(char *s);
 int			is_pipe_or_redir(char *c);
 int			from_pipe(t_treenode *node);
 int			to_pipe(t_treenode *node);
 int			takes_part_in_pipe(t_treenode *node);
-int			is_file(char *c);
-int			from_file(t_treenode *node);
-int			to_file(t_treenode *node);
+int			is_to_out_redir(char *c);
+int			is_from_in_redir(char *c);
+int			is_out_file(t_treenode *node);
+int			is_in_file(t_treenode *node);
+int			sends_to_out_file(t_treenode *node);
+int			reads_from_in_file(t_treenode *node);
 int			takes_part_in_file(t_treenode *node);
 #endif
